LPD-80236 Add useColorPicker hook
Combine all the logic of the color picker into a hook so it can be reused.
