+++ modules/apps/frontend-js/frontend-js-clay-web/clay/clay-color-picker/src/Editor.tsx
@@ -40,7 +40,7 @@ export function useEditor(
 ) {
 	const [state, dispatch] = useReducer(reducer, {}, () => {
 		const index = colors.findIndex(
-			(color) => color.toUpperCase() === value.toUpperCase()
+			(color) => color?.toUpperCase() === value?.toUpperCase()
 		);
 
 		return {
